# next/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/register": {
      "post": {
        "operationId": "api_auth_register_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponseDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "api"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "api_auth_login_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "api"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "api_auth_logout_create",
        "parameters": [],
        "responses": {
          "205": {
            "description": ""
          }
        },
        "security": [
          {
            "jwtAuth": []
          }
        ],
        "summary": "",
        "tags": [
          "api"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "api_auth_me_retrieve",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwtAuth": []
          }
        ],
        "summary": "",
        "tags": [
          "api"
        ]
      }
    },
    "/auth/token/refresh": {
      "post": {
        "operationId": "api_auth_token_refresh_create",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "api"
        ]
      }
    },
    "/users": {
      "get": {
        "operationId": "api_users_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "exclusiveMinimum": false,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "name": "ordering",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "name",
                "-name",
                "created_at",
                "-created_at",
                "updated_at",
                "-updated_at"
              ]
            }
          },
          {
            "name": "created_at__gte",
            "required": false,
            "in": "query",
            "x-nestjs_zod-empty-type": true,
            "schema": {}
          },
          {
            "name": "created_at__lte",
            "required": false,
            "in": "query",
            "x-nestjs_zod-empty-type": true,
            "schema": {}
          },
          {
            "name": "updated_at__gte",
            "required": false,
            "in": "query",
            "x-nestjs_zod-empty-type": true,
            "schema": {}
          },
          {
            "name": "updated_at__lte",
            "required": false,
            "in": "query",
            "x-nestjs_zod-empty-type": true,
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedUserResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwtAuth": []
          }
        ],
        "summary": "",
        "tags": [
          "api"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "api_users_retrieve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwtAuth": []
          }
        ],
        "summary": "",
        "tags": [
          "api"
        ]
      },
      "patch": {
        "operationId": "api_users_partial_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwtAuth": []
          }
        ],
        "summary": "",
        "tags": [
          "api"
        ]
      }
    },
    "/todos": {
      "get": {
        "operationId": "api_todos_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "exclusiveMinimum": false,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "name": "ordering",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "title",
                "-title",
                "created_at",
                "-created_at",
                "updated_at",
                "-updated_at"
              ]
            }
          },
          {
            "name": "created_at__gte",
            "required": false,
            "in": "query",
            "x-nestjs_zod-empty-type": true,
            "schema": {}
          },
          {
            "name": "created_at__lte",
            "required": false,
            "in": "query",
            "x-nestjs_zod-empty-type": true,
            "schema": {}
          },
          {
            "name": "updated_at__gte",
            "required": false,
            "in": "query",
            "x-nestjs_zod-empty-type": true,
            "schema": {}
          },
          {
            "name": "updated_at__lte",
            "required": false,
            "in": "query",
            "x-nestjs_zod-empty-type": true,
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedTodoResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwtAuth": []
          }
        ],
        "summary": "",
        "tags": [
          "api"
        ]
      },
      "post": {
        "operationId": "api_todos_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoCreateRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwtAuth": []
          }
        ],
        "summary": "",
        "tags": [
          "api"
        ]
      }
    },
    "/todos/{id}": {
      "get": {
        "operationId": "api_todos_retrieve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwtAuth": []
          }
        ],
        "summary": "",
        "tags": [
          "api"
        ]
      },
      "patch": {
        "operationId": "api_todos_partial_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoUpdateRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwtAuth": []
          }
        ],
        "summary": "",
        "tags": [
          "api"
        ]
      },
      "delete": {
        "operationId": "api_todos_destroy",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "jwtAuth": []
          }
        ],
        "summary": "",
        "tags": [
          "api"
        ]
      }
    }
  },
  "info": {
    "title": "OpenAPI Schema",
    "description": "Prisma-driven NestJS API schema",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "jwtAuth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token"
      }
    },
    "schemas": {
      "RegisterRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "email",
          "password",
          "name"
        ]
      },
      "RegisterResponseDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "integer"
          }
        },
        "required": [
          "email",
          "name",
          "id"
        ]
      },
      "LoginRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "required": [
          "email"
        ]
      },
      "CurrentUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "email": {
            "type": "string"
          },
          "avatar": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "birthday": {
            "nullable": true,
            "type": "",
            "x-nestjs_zod-empty-type": true
          },
          "is_active": {
            "type": "boolean"
          },
          "is_staff": {
            "type": "boolean"
          },
          "is_email_verified": {
            "type": "boolean"
          },
          "is_phone_verified": {
            "type": "boolean"
          },
          "created_at": {
            "type": "",
            "x-nestjs_zod-empty-type": true
          },
          "updated_at": {
            "type": "",
            "x-nestjs_zod-empty-type": true
          },
          "groups": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer"
                },
                "name": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "name"
              ]
            }
          }
        },
        "required": [
          "id",
          "email",
          "avatar",
          "name",
          "phone",
          "birthday",
          "is_active",
          "is_staff",
          "is_email_verified",
          "is_phone_verified",
          "created_at",
          "updated_at",
          "groups"
        ]
      },
      "PaginatedUserResponseDto": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer"
          },
          "next": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "previous": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "results": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer"
                },
                "email": {
                  "type": "string"
                },
                "avatar": {
                  "type": "string",
                  "nullable": true
                },
                "name": {
                  "type": "string",
                  "nullable": true
                },
                "phone": {
                  "type": "string",
                  "nullable": true
                },
                "birthday": {
                  "nullable": true
                },
                "is_active": {
                  "type": "boolean"
                },
                "is_staff": {
                  "type": "boolean"
                },
                "is_email_verified": {
                  "type": "boolean"
                },
                "is_phone_verified": {
                  "type": "boolean"
                },
                "created_at": {},
                "updated_at": {},
                "groups": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "name": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "name"
                    ]
                  }
                }
              },
              "required": [
                "id",
                "email",
                "avatar",
                "name",
                "phone",
                "birthday",
                "is_active",
                "is_staff",
                "is_email_verified",
                "is_phone_verified",
                "created_at",
                "updated_at",
                "groups"
              ]
            }
          }
        },
        "required": [
          "count",
          "next",
          "previous",
          "results"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "email": {
            "type": "string"
          },
          "avatar": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "birthday": {
            "nullable": true,
            "type": "",
            "x-nestjs_zod-empty-type": true
          },
          "is_active": {
            "type": "boolean"
          },
          "is_staff": {
            "type": "boolean"
          },
          "is_email_verified": {
            "type": "boolean"
          },
          "is_phone_verified": {
            "type": "boolean"
          },
          "created_at": {
            "type": "",
            "x-nestjs_zod-empty-type": true
          },
          "updated_at": {
            "type": "",
            "x-nestjs_zod-empty-type": true
          },
          "groups": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer"
                },
                "name": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "name"
              ]
            }
          }
        },
        "required": [
          "id",
          "email",
          "avatar",
          "name",
          "phone",
          "birthday",
          "is_active",
          "is_staff",
          "is_email_verified",
          "is_phone_verified",
          "created_at",
          "updated_at",
          "groups"
        ]
      },
      "UserUpdateRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "avatar": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "birthday": {
            "nullable": true,
            "type": "",
            "x-nestjs_zod-empty-type": true
          }
        }
      },
      "PaginatedTodoResponseDto": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer"
          },
          "next": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "previous": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "results": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "title": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "is_complete": {
                  "type": "boolean"
                },
                "created_at": {},
                "updated_at": {},
                "owner": {
                  "type": "integer"
                },
                "owner_email": {
                  "type": "string",
                  "format": "email"
                }
              },
              "required": [
                "id",
                "title",
                "description",
                "is_complete",
                "created_at",
                "updated_at",
                "owner",
                "owner_email"
              ]
            }
          }
        },
        "required": [
          "count",
          "next",
          "previous",
          "results"
        ]
      },
      "TodoCreateRequestDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "is_complete": {
            "type": "boolean"
          }
        },
        "required": [
          "title"
        ]
      },
      "TodoResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "is_complete": {
            "type": "boolean"
          },
          "created_at": {
            "type": "",
            "x-nestjs_zod-empty-type": true
          },
          "updated_at": {
            "type": "",
            "x-nestjs_zod-empty-type": true
          },
          "owner": {
            "type": "integer"
          },
          "owner_email": {
            "type": "string",
            "format": "email"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "is_complete",
          "created_at",
          "updated_at",
          "owner",
          "owner_email"
        ]
      },
      "TodoUpdateRequestDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "is_complete": {
            "type": "boolean"
          }
        }
      }
    }
  }
}